Move WebTransport stuff out
This means that if you want webtransport, you need to import the trait.

This file includes both the client and server pieces because it isn't
that much code.  The net effect is a reduction in code size for other
files.
