+++ neqo-http3/src/features/extended_connect/tests/webtransport/streams.rs
@@ -9,6 +9,7 @@ use neqo_transport::StreamType;
 use crate::{
     Error,
     features::extended_connect::{CloseReason, tests::webtransport::WtTest},
+    webtransport::WebTransport as _,
 };
 
 #[test]
