+++ neqo-http3/src/lib.rs
@@ -24,8 +24,8 @@ implementation is not meant to be used in production and its only purpose is to 
 of the client-side code.
 
 __`WebTransport`__
-([draft version 2](https://datatracker.ietf.org/doc/html/draft-vvv-webtransport-http3-02)) is
-supported and can be enabled using [`Http3Parameters`](struct.Http3Parameters.html).
+WebTransport is supported
+and can be enabled using [`Http3Parameters`](struct.Http3Parameters.html).
 
 ## Interaction with an application
 
@@ -161,6 +161,7 @@ mod server_connection_events;
 mod server_events;
 mod settings;
 mod stream_type_reader;
+pub mod webtransport;
 
 use std::{cell::RefCell, fmt::Debug, rc::Rc, time::Instant};
 
@@ -180,7 +181,6 @@ pub use push_id::PushId;
 pub use server::Http3Server;
 pub use server_events::{
     ConnectUdpRequest, ConnectUdpServerEvent, Http3OrWebTransportStream, Http3ServerEvent,
-    WebTransportRequest, WebTransportServerEvent,
 };
 #[cfg(fuzzing)]
 pub use settings::HSettings;
