+++ neqo-http3/src/server_events.rs
@@ -13,19 +13,15 @@ use std::{
     time::Instant,
 };
 
-use neqo_common::{Bytes, Encoder, Header, qdebug};
-use neqo_transport::{
-    AppError, Connection, DatagramTracking, StreamId, StreamType, server::ConnectionRef,
-};
+use neqo_common::{Bytes, Header, qdebug};
+use neqo_transport::{AppError, Connection, DatagramTracking, StreamId, server::ConnectionRef};
 
 use crate::{
-    Error, Http3StreamInfo, Http3StreamType, Priority, Res,
+    Http3StreamInfo, Http3StreamType, Priority, Res,
     connection::{Http3State, SessionAcceptAction},
     connection_server::Http3ServerHandler,
-    features::{
-        extended_connect,
-        extended_connect::webtransport_session::WebTransportExportKeyingMaterial as _,
-    },
+    features::extended_connect,
+    webtransport::WebTransportServerEvent,
 };
 
 #[derive(Debug, Clone)]
@@ -237,176 +233,6 @@ impl PartialEq for Http3OrWebTransportStream {
 impl Eq for Http3OrWebTransportStream {}
 
 #[derive(Debug, Clone)]
-pub struct WebTransportRequest {
-    stream_handler: StreamHandler,
-}
-
-impl Display for WebTransportRequest {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "WebTransport session {}", self.stream_handler)
-    }
-}
-
-impl WebTransportRequest {
-    pub(crate) const fn new(
-        conn: ConnectionRef,
-        handler: Rc<RefCell<Http3ServerHandler>>,
-        stream_id: StreamId,
-    ) -> Self {
-        Self {
-            stream_handler: StreamHandler {
-                conn,
-                handler,
-                stream_info: Http3StreamInfo::new(stream_id, Http3StreamType::Http),
-            },
-        }
-    }
-
-    #[must_use]
-    pub fn state(&self) -> Http3State {
-        self.stream_handler.handler.borrow().state()
-    }
-
-    /// Respond to a `WebTransport` session request.
-    ///
-    /// # Errors
-    ///
-    /// It may return `InvalidStreamId` if a stream does not exist anymore.
-    pub fn response(&self, accept: &SessionAcceptAction, now: Instant) -> Res<()> {
-        qdebug!("[{self}] Set a response for a WebTransport session");
-        self.stream_handler
-            .handler
-            .borrow_mut()
-            .webtransport_session_accept(
-                &mut self.stream_handler.conn.borrow_mut(),
-                self.stream_handler.stream_info.stream_id(),
-                accept,
-                now,
-            )
-    }
-
-    /// # Errors
-    ///
-    /// It may return `InvalidStreamId` if a stream does not exist anymore.
-    /// Also return an error if the stream was closed on the transport layer,
-    /// but that information is not yet consumed on the  http/3 layer.
-    pub fn close_session(&self, error: u32, message: &str, now: Instant) -> Res<()> {
-        self.stream_handler
-            .handler
-            .borrow_mut()
-            .webtransport_close_session(
-                &mut self.stream_handler.conn.borrow_mut(),
-                self.stream_handler.stream_info.stream_id(),
-                error,
-                message,
-                now,
-            )
-    }
-
-    #[must_use]
-    pub const fn stream_id(&self) -> StreamId {
-        self.stream_handler.stream_id()
-    }
-
-    /// Create `WebTransport` stream.
-    ///
-    /// # Errors
-    ///
-    /// It may return `InvalidStreamId` if a stream does not exist anymore.
-    pub fn create_stream(&self, stream_type: StreamType) -> Res<Http3OrWebTransportStream> {
-        let session_id = self.stream_handler.stream_id();
-        let id = self
-            .stream_handler
-            .handler
-            .borrow_mut()
-            .webtransport_create_stream(
-                &mut self.stream_handler.conn.borrow_mut(),
-                session_id,
-                stream_type,
-            )?;
-
-        Ok(Http3OrWebTransportStream::new(
-            self.stream_handler.conn.clone(),
-            Rc::clone(&self.stream_handler.handler),
-            Http3StreamInfo::new(id, Http3StreamType::WebTransport(session_id)),
-        ))
-    }
-
-    /// Send `WebTransport` datagram.
-    ///
-    /// # Errors
-    ///
-    /// It may return `InvalidStreamId` if a stream does not exist anymore.
-    /// The function returns `TooMuchData` if the supply buffer is bigger than
-    /// the allowed remote datagram size.
-    pub fn send_datagram<I: Into<DatagramTracking>>(
-        &self,
-        buf: &[u8],
-        id: I,
-        now: Instant,
-    ) -> Res<()> {
-        let session_id = self.stream_handler.stream_id();
-        self.stream_handler
-            .handler
-            .borrow_mut()
-            .webtransport_send_datagram(
-                &mut self.stream_handler.conn.borrow_mut(),
-                session_id,
-                buf,
-                id,
-                now,
-            )
-    }
-
-    // TODO: Currently not called in neqo or gecko. It should likely be called at least from gecko.
-    #[must_use]
-    pub fn remote_datagram_size(&self) -> u64 {
-        self.stream_handler.conn.borrow().remote_datagram_size()
-    }
-
-    /// Returns the current max size of a datagram that can fit into a packet.
-    /// The value will change over time depending on the encoded size of the
-    /// packet number, ack frames, etc.
-    ///
-    /// # Errors
-    ///
-    /// The function returns `NotAvailable` if datagrams are not enabled.
-    ///
-    /// # Panics
-    ///
-    /// This cannot panic. The max varint length is 8.
-    pub fn max_datagram_size(&self) -> Res<u64> {
-        let max_size = self.stream_handler.conn.borrow().max_datagram_size()?;
-        Ok(max_size
-            - u64::try_from(Encoder::varint_len(
-                self.stream_handler.stream_id().as_u64(),
-            ))
-            .map_err(|_| Error::Internal)?)
-    }
-
-    /// Export keying material for this WebTransport session
-    /// (draft-ietf-webtrans-http3 §4.8).
-    ///
-    /// # Errors
-    ///
-    /// Returns `Error::InvalidStreamId` if the session is no longer active,
-    /// `Error::InvalidInput` if `out` is empty or `label`/`context`
-    /// exceed 255 bytes, or `Error::Transport` if the connection is not ready
-    /// or the TLS export fails.
-    pub fn export_keying_material(&self, label: &[u8], context: &[u8], out: &mut [u8]) -> Res<()> {
-        let session_id = self.stream_handler.stream_id();
-        self.stream_handler
-            .handler
-            .borrow()
-            .validate_extended_connect_session(session_id)?;
-        self.stream_handler
-            .conn
-            .borrow()
-            .webtransport_export_keying_material(session_id, label, context, out)
-    }
-}
-
-#[derive(Debug, Clone)]
 pub struct ConnectUdpRequest {
     stream_handler: StreamHandler,
 }
@@ -523,31 +349,6 @@ impl ConnectUdpRequest {
     }
 }
 
-impl Deref for WebTransportRequest {
-    type Target = StreamHandler;
-    fn deref(&self) -> &Self::Target {
-        &self.stream_handler
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum WebTransportServerEvent {
-    NewSession {
-        session: WebTransportRequest,
-        headers: Vec<Header>,
-    },
-    SessionClosed {
-        session: WebTransportRequest,
-        reason: extended_connect::session::CloseReason,
-        headers: Option<Vec<Header>>,
-    },
-    NewStream(Http3OrWebTransportStream),
-    Datagram {
-        session: WebTransportRequest,
-        datagram: Bytes,
-    },
-}
-
 #[derive(Debug, Clone)]
 pub enum ConnectUdpServerEvent {
     NewSession {
@@ -610,7 +411,7 @@ pub struct Http3ServerEvents {
 }
 
 impl Http3ServerEvents {
-    fn insert(&self, event: Http3ServerEvent) {
+    pub(crate) fn insert(&self, event: Http3ServerEvent) {
         self.events.borrow_mut().push_back(event);
     }
 
@@ -708,34 +509,9 @@ impl Http3ServerEvents {
         });
     }
 
-    pub(crate) fn webtransport_new_session(
-        &self,
-        session: WebTransportRequest,
-        headers: Vec<Header>,
-    ) {
-        self.insert(Http3ServerEvent::WebTransport(
-            WebTransportServerEvent::NewSession { session, headers },
-        ));
-    }
-
     pub(crate) fn connect_udp_new_session(&self, session: ConnectUdpRequest, headers: Vec<Header>) {
         self.insert(Http3ServerEvent::ConnectUdp(
             ConnectUdpServerEvent::NewSession { session, headers },
-        ));
-    }
-
-    pub(crate) fn webtransport_session_closed(
-        &self,
-        session: WebTransportRequest,
-        reason: extended_connect::session::CloseReason,
-        headers: Option<Vec<Header>>,
-    ) {
-        self.insert(Http3ServerEvent::WebTransport(
-            WebTransportServerEvent::SessionClosed {
-                session,
-                reason,
-                headers,
-            },
         ));
     }
 
@@ -754,17 +530,6 @@ impl Http3ServerEvents {
         ));
     }
 
-    pub(crate) fn webtransport_new_stream(&self, stream: Http3OrWebTransportStream) {
-        self.insert(Http3ServerEvent::WebTransport(
-            WebTransportServerEvent::NewStream(stream),
-        ));
-    }
-
-    pub(crate) fn webtransport_datagram(&self, session: WebTransportRequest, datagram: Bytes) {
-        self.insert(Http3ServerEvent::WebTransport(
-            WebTransportServerEvent::Datagram { session, datagram },
-        ));
-    }
     pub(crate) fn connect_udp_datagram(&self, session: ConnectUdpRequest, datagram: Bytes) {
         self.insert(Http3ServerEvent::ConnectUdp(
             ConnectUdpServerEvent::Datagram { session, datagram },
