+++ neqo-http3/tests/webtransport.rs
@@ -11,8 +11,8 @@ use std::{cell::RefCell, rc::Rc};
 use neqo_common::{event::Provider as _, header::HeadersExt as _};
 use neqo_http3::{
     Http3Client, Http3ClientEvent, Http3OrWebTransportStream, Http3Parameters, Http3Server,
-    Http3ServerEvent, Http3State, SessionAcceptAction, WebTransportEvent, WebTransportRequest,
-    WebTransportServerEvent,
+    Http3ServerEvent, Http3State, SessionAcceptAction, WebTransportEvent,
+    webtransport::{WebTransport as _, WebTransportRequest, WebTransportServerEvent},
 };
 use neqo_transport::{ConnectionParameters, StreamId, StreamType};
 use nss::AuthenticationStatus;
